+++ ledger/ledger.go
@@ -500,13 +500,6 @@ func (l *Ledger) Timestamp(r basics.Round) (int64, error) {
 	return l.time.timestamp(r)
 }
 
-// AllBalances returns a map of every account balance as of round rnd.
-func (l *Ledger) AllBalances(rnd basics.Round) (map[basics.Address]basics.AccountData, error) {
-	l.trackerMu.RLock()
-	defer l.trackerMu.RUnlock()
-	return l.accts.allBalances(rnd)
-}
-
 // GenesisHash returns the genesis hash for this ledger.
 func (l *Ledger) GenesisHash() crypto.Digest {
 	return l.genesisHash
